+++ certbot/tests/client_test.py
@@ -89,7 +89,7 @@ class RegisterTest(test_util.ConfigTestCase):
         with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_client:
             mock_client.new_account_and_tos().terms_of_service = "http://tos"
             mock_client().external_account_required.side_effect = self._false_mock
-            with mock.patch("certbot.eff.handle_subscription") as mock_handle:
+            with mock.patch("certbot._internal.eff.handle_subscription") as mock_handle:
                 with mock.patch("certbot._internal.account.report_new_account"):
                     mock_client().new_account_and_tos.side_effect = errors.Error
                     self.assertRaises(errors.Error, self._call)
@@ -103,7 +103,7 @@ class RegisterTest(test_util.ConfigTestCase):
         with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_client:
             mock_client().external_account_required.side_effect = self._false_mock
             with mock.patch("certbot._internal.account.report_new_account"):
-                with mock.patch("certbot.eff.handle_subscription"):
+                with mock.patch("certbot._internal.eff.handle_subscription"):
                     self._call()
 
     @mock.patch("certbot._internal.account.report_new_account")
@@ -115,7 +115,7 @@ class RegisterTest(test_util.ConfigTestCase):
         mx_err = messages.Error.with_code('invalidContact', detail=msg)
         with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_client:
             mock_client().external_account_required.side_effect = self._false_mock
-            with mock.patch("certbot.eff.handle_subscription") as mock_handle:
+            with mock.patch("certbot._internal.eff.handle_subscription") as mock_handle:
                 mock_client().new_account_and_tos.side_effect = [mx_err, mock.MagicMock()]
                 self._call()
                 self.assertEqual(mock_get_email.call_count, 1)
@@ -129,7 +129,7 @@ class RegisterTest(test_util.ConfigTestCase):
         mx_err = messages.Error.with_code('invalidContact', detail=msg)
         with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_client:
             mock_client().external_account_required.side_effect = self._false_mock
-            with mock.patch("certbot.eff.handle_subscription"):
+            with mock.patch("certbot._internal.eff.handle_subscription"):
                 mock_client().new_account_and_tos.side_effect = [mx_err, mock.MagicMock()]
                 self.assertRaises(errors.Error, self._call)
 
@@ -139,7 +139,7 @@ class RegisterTest(test_util.ConfigTestCase):
 
     @mock.patch("certbot._internal.client.logger")
     def test_without_email(self, mock_logger):
-        with mock.patch("certbot.eff.handle_subscription") as mock_handle:
+        with mock.patch("certbot._internal.eff.handle_subscription") as mock_handle:
             with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_clnt:
                 mock_clnt().external_account_required.side_effect = self._false_mock
                 with mock.patch("certbot._internal.account.report_new_account"):
@@ -156,7 +156,7 @@ class RegisterTest(test_util.ConfigTestCase):
         """Tests dry-run for no staging account, expect account created with no email"""
         with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_client:
             mock_client().external_account_required.side_effect = self._false_mock
-            with mock.patch("certbot.eff.handle_subscription"):
+            with mock.patch("certbot._internal.eff.handle_subscription"):
                 with mock.patch("certbot._internal.account.report_new_account"):
                     self.config.dry_run = True
                     self._call()
@@ -171,7 +171,7 @@ class RegisterTest(test_util.ConfigTestCase):
                 side_effect=self._new_acct_dir_mock
             )
             mock_client().external_account_required.side_effect = self._false_mock
-            with mock.patch("certbot.eff.handle_subscription"):
+            with mock.patch("certbot._internal.eff.handle_subscription"):
                 target = "certbot._internal.client.messages.ExternalAccountBinding.from_data"
                 with mock.patch(target) as mock_eab_from_data:
                     self.config.eab_kid = "test-kid"
@@ -183,7 +183,7 @@ class RegisterTest(test_util.ConfigTestCase):
     def test_without_eab_arguments(self):
         with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_client:
             mock_client().external_account_required.side_effect = self._false_mock
-            with mock.patch("certbot.eff.handle_subscription"):
+            with mock.patch("certbot._internal.eff.handle_subscription"):
                 target = "certbot._internal.client.messages.ExternalAccountBinding.from_data"
                 with mock.patch(target) as mock_eab_from_data:
                     self.config.eab_kid = None
@@ -196,7 +196,7 @@ class RegisterTest(test_util.ConfigTestCase):
         with mock.patch("certbot._internal.client.acme_client.BackwardsCompatibleClientV2") as mock_client:
             mock_client().client.net.key.public_key = mock.Mock(side_effect=self._public_key_mock)
             mock_client().external_account_required.side_effect = self._true_mock
-            with mock.patch("certbot.eff.handle_subscription"):
+            with mock.patch("certbot._internal.eff.handle_subscription"):
                 with mock.patch("certbot._internal.client.messages.ExternalAccountBinding.from_data"):
                     self.config.eab_kid = None
                     self.config.eab_hmac_key = None
@@ -212,7 +212,7 @@ class RegisterTest(test_util.ConfigTestCase):
                 side_effect=self._new_acct_dir_mock
             )
             mock_client().external_account_required.side_effect = self._false_mock
-            with mock.patch("certbot.eff.handle_subscription") as mock_handle:
+            with mock.patch("certbot._internal.eff.handle_subscription") as mock_handle:
                 mock_client().new_account_and_tos.side_effect = [mx_err, mock.MagicMock()]
                 self.assertRaises(messages.Error, self._call)
         self.assertFalse(mock_handle.called)
